+++ lib/puppet/type/wls_setting/use_default_value_when_empty.rb
@@ -1,0 +1,14 @@
+newproperty(:use_default_value_when_empty) do
+  include EasyType
+
+  desc 'Use default value when attribute value is not provided'
+
+  newvalues(:true, :false)
+
+  defaultto :true
+
+  to_translate_to_resource do |raw_resource|
+    raw_resource[self.name]
+  end
+
+end