+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - wls_resource_group new MT type but only for WebLogic 12.2.1
 - wls_domain_partition new MT type but only for WebLogic 12.2.1
 - wls_domain_partition_resource_group new MT type but only for WebLogic 12.2.1
+- wls_setting new attribute use_default_value_when_empty, when you want to make sure the wls type properties will set its default mbean values when it is not provided by your puppet configuration.
 
 ## 1.0.47
 - heavy WLST refactoring for all the wls types, less and more simple code, auto detect the right type, better fault handling